+++ confortable.js
@@ -1,46 +1,41 @@
-var fs   = require('fs')
-  , path = require('path')
-  , existsSync = fs.existsSync || path.existsSync;
+var exists = require('fs').existsSync
+  , path = require('path');
+
+var withinHome = function (dir) {
+  return (path.relative(process.env.HOME, dir).slice(0, 2) !== '..');
+};
 
 var findFromStart = function (name, start) {
   start = start || process.cwd();
   var relative = path.relative(process.env.HOME, start)
     , noRelation = (relative === start) // i.e. different drive or undefined HOME
-    , isAbove = (relative.slice(0, 2) === '..')
-    , cfg;
+    , isAbove = (relative.slice(0, 2) === '..');
 
   if (noRelation || isAbove) {
-    // start is outside home, check start only
-    cfg = path.join(start, name);
-    if (existsSync(cfg)) {
-      return cfg;
+    // start is outside HOME: check start and HOME only
+    var startCfg = path.join(start, name)
+      , homeCfg = path.join(process.env.HOME, name);
+
+    if (exists(startCfg)) {
+      return startCfg;
     }
-    else {
-      cfg = path.join(process.env.HOME, name);
-      return existsSync(cfg) ? cfg : null;
+    if (exists(homeCfg)) {
+      return homeCfg;
     }
+    return;
   }
-  else {
-    // start is somewhere under HOME start, so start there and go up until we hit HOME
-    var dir = start;
-    while (true) {
-      relative = path.relative(process.env.HOME, dir);
-      if (relative.slice(0, 2) === '..') {
-        break; // dir is above HOME
-      }
 
-      cfg = path.join(dir, name);
-      if (existsSync(cfg)) {
-        return cfg;
-      }
-
-      // windows users can put a drive root to home in which case adding '..' does nothing
-      if (path.join(dir, '..') === path.normalize(dir)) {
-        break; // next iteration's dir is root, and nothing's there, prevent infinite loop
-      }
-      dir = path.join(dir, '..');
+  // start is under HOME, move up the directory tree until we hit HOME
+  for (var dir = start; withinHome(dir) ; dir = path.join(dir, '..')) {
+    var currCfg = path.join(dir, name);
+    if (exists(currCfg)) {
+      return currCfg;
     }
-    return null;
+
+    // windows $HOME defined at drive root => joining on '..' does nothing
+    if (path.join(dir, '..') === path.normalize(dir)) {
+      return; // at windows root, we found nothing. dont loop forever
+    }
   }
 };
 
@@ -51,5 +46,7 @@ module.exports = function (name, start, fallbackDir) {
     return result;
   }
   var fbPath = path.join(fallbackDir, name);
-  return (existsSync(fbPath)) ? fbPath : null;
+  if (exists(fbPath)) {
+    return fbPath;
+  }
 };