+++ test/basic.js
@@ -1,8 +1,7 @@
 var conf = require('../')
   , fs   = require('fs')
   , test = require('tap').test
-  , path = require('path')
-  , join = path.join;
+  , join = require('path').join;
 
 
 test("location", function (t) {
@@ -30,4 +29,23 @@ test("location", function (t) {
   t.end();
 });
 
+test("home border", function (t) {
+  var cwd = process.cwd();
+  var name = '.confconf';
 
+  process.env.HOME = join(cwd, name); // pretend we're in HOME
+
+  var files = [
+    join(cwd, '..', name)
+  , join(cwd, name)
+  ]
+
+  fs.writeFileSync(files[0], "{}");
+  t.equal(conf(name), undefined, "conf does not find above HOME");
+
+  fs.writeFileSync(files[1], "{}");
+  t.equal(conf(name), files[1], "conf finds in cwd");
+
+  files.map(fs.unlinkSync);
+  t.end();
+});